+++ ChangeLog.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Filesystem::link()
 - Filesystem::symlink()
 - Filesystem::rename()
+- Filesystem::truncate()
 
 ## [0.0.3] - 2024-10-25